tools: Fix compilation issues on 32 bit machines
size_t is 32 bits on those machines, failing as an argument to
printf. Some variables in oping and operf were changed to uint32_t and
uint64_t to avoid issues.

+++ src/tools/operf/operf.c
@@ -44,8 +44,8 @@ struct c {
         bool   sleep;
         int    duration;
 
-        size_t sent;
-        size_t rcvd;
+        uint64_t sent;
+        uint64_t rcvd;
 
         flow_set_t * flows;
         fqueue_t *   fq;
